Fixes headers when there is no login/passcode
In the current implementation, if the user does not define a login and
passcode, then the library sets their values to '' and the headers that
are sent to the STOMP server are the following:

```
CONNECT
accept-version:1.0,1.1
host:stomp-server.example.com
heart-beat:0,0
login:
passcode:
```

When this is tested against ActiveMQ (version 5.8.0), the connection
fails with the following error:

```
ERROR
content-type:text/plain
message:User name [] or password is invalid.
```

According to the STOMP protocol specification, the clients MAY set the
login and passcode headers. It seems that when the user does not define
a login and passcode, then the CONNECT frame must not include these
headers.

This commit changes the existing behaviour and when the user does not
define a login and a passcode, then the library does not include these
headers in the CONNECT frame.